requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
groq
google-cloud-vision
PyPDF2
python-docx
Pillow
PyMuPDF
requests
beautifulsoup4
SpeechRecognition
httpx
pydub